refactor: MemberRole enum 파일 위치 users -> projects로 변경
