+++ src/main/java/com/growlog/webide/domain/permissions/service/ProjectPermissionService.java
@@ -3,9 +3,9 @@ package com.growlog.webide.domain.permissions.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.growlog.webide.domain.projects.entity.MemberRole;
 import com.growlog.webide.domain.projects.entity.Project;
 import com.growlog.webide.domain.projects.repository.ProjectMemberRepository;
-import com.growlog.webide.domain.users.entity.MemberRole;
 import com.growlog.webide.global.common.exception.CustomException;
 import com.growlog.webide.global.common.exception.ErrorCode;
 import com.growlog.webide.global.security.UserPrincipal;
@@ -19,16 +19,14 @@ public class ProjectPermissionService {
 
 	@Transactional(readOnly = true)
 	public MemberRole getMyRole(Long projectId, UserPrincipal userPrincipal) {
-		return projectMemberRepository
-			.findByProject_IdAndUser_UserId(projectId, userPrincipal.getUserId())
+		return projectMemberRepository.findByProject_IdAndUser_UserId(projectId, userPrincipal.getUserId())
 			.map(pm -> pm.getRole())
 			.orElseThrow(() -> new CustomException(ErrorCode.NOT_A_MEMBER));
 	}
 
 	@Transactional(readOnly = true)
 	public void checkReadAccess(Project project, Long userId) {
-		MemberRole role = projectMemberRepository
-			.findByProject_IdAndUser_UserId(project.getId(), userId)
+		MemberRole role = projectMemberRepository.findByProject_IdAndUser_UserId(project.getId(), userId)
 			.map(pm -> pm.getRole())
 			.orElseThrow(() -> new CustomException(ErrorCode.NOT_A_MEMBER));
