+++ src/main/java/com/growlog/webide/domain/projects/entity/MemberRole.java
@@ -1,0 +1,7 @@
+package com.growlog.webide.domain.projects.entity;
+
+public enum MemberRole {
+	OWNER,
+	READ,
+	WRITE
+}
