--- src/main/java/com/growlog/webide/domain/users/entity/MemberRole.java
@@ -1,7 +1,0 @@
-package com.growlog.webide.domain.users.entity;
-
-public enum MemberRole {
-	OWNER,
-	READ,
-	WRITE
-}
